requirements.txt
autopep8==1.6.0
boto3==1.20.15
botocore==1.23.15
certifi==2021.10.8
chardet==3.0.4
click==8.0.3
crypto==1.4.1
Flask==2.0.2
greenlet==1.1.2
idna==2.8
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
Naked==0.1.31
phonenumbers==8.12.37
pycodestyle==2.8.0
PyMySQL==1.0.2
python-dateutil==2.8.2
PyYAML==6.0
requests==2.22.0
s3transfer==0.5.0
shellescape==3.8.1
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
urllib3==1.25.11
Werkzeug==2.0.2
